requirements.txt
certifi
chardet
future
idna
oauthlib
python-twitter
PyYAML
requests
requests-oauthlib
urllib3
